实现点击收藏功能；优化cookie登录逻辑(header)，添加房车页面(待完善)

+++ springboot-01-helloworld/src/main/java/com/zhao/pojo/User.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @AllArgsConstructor
 @Data
 public class User {
-    public long user_id;
+    public int user_id;
     public String user_name;
     public String user_password;
     public String user_nickname;
