+++ springboot-01-helloworld/src/main/java/com/zhao/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
 
     User findUserByNickName(String nickname);
 
-    void changeUserImageUrl(String imageUrl, long user_id);
+    void changeUserImageUrl(String imageUrl, int user_id);
 
     boolean ifUserExist(String username);
